refactor: 💡 QueryRunnerを分離する
